Extract out loadPageNoOkta
